Moved blueprints into separate folder
