+++ dashboard/blueprints/scans/templates/snips/dcm2niix_details.html
