+++ dashboard/blueprints/scans/templates/snips/header.html
