+++ dashboard/blueprints/scans/templates/snips/header_diffs.html
