+++ dashboard/blueprints/scans/templates/snips/papaya_nav.html
