+++ dashboard/blueprints/timepoints/templates/collapsible_section.html
