+++ dashboard/blueprints/timepoints/templates/modals/delete_data.html
