+++ dashboard/blueprints/timepoints/templates/sessions/scan_table.html
