+++ dashboard/blueprints/users/templates/account_request.html
