+++ dashboard/blueprints/users/templates/snips/permissions.html
