+++ dashboard/blueprints/users/views.py
@@ -6,8 +6,8 @@ from dashboard import lm
 from . import user_bp
 from .utils import get_user_form, parse_enabled_sites
 from .forms import UserForm
-from ..models import User, AccountRequest
-from ..utils import report_form_errors, dashboard_admin_required
+from ...models import User, AccountRequest
+from ...utils import report_form_errors, dashboard_admin_required
 
 
 @lm.user_loader
